Error Handling

+++ src/main/java/com/example/boxticketingwebapi/service/AccountService.java
@@ -1,6 +1,7 @@
 package com.example.boxticketingwebapi.service;
 
 
+import com.example.boxticketingwebapi.DataNotFoundException;
 import com.example.boxticketingwebapi.model.AccountModel;
 import com.example.boxticketingwebapi.repo.IAccountRepo;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +19,10 @@ public class AccountService {
     }
 
     public String deleteAccount(Integer id){
+        AccountModel account = this.accountRepo.findById(id).orElse(null);
+        if(account == null) {
+            throw new DataNotFoundException("Account with id:" + id + " doesn't exists!");
+        }
         accountRepo.deleteById(id);
         return "Account Deleted !";
     }
